Move affect perception over to perception side
